Add SchemaBuilderMaterials
