+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/metadata/schema/builder/loader/dialect/impl/SQL92DatabaseMetaDataDialectHandlerTest.java
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.metadata.schema.builder.physical.dialect.impl;
+package org.apache.shardingsphere.infra.metadata.schema.builder.loader.dialect.impl;
 
-import org.apache.shardingsphere.infra.database.type.dialect.H2DatabaseType;
-import org.apache.shardingsphere.infra.metadata.schema.builder.physical.dialect.AbstractDatabaseMetaDataDialectHandlerTest;
+import org.apache.shardingsphere.infra.database.type.dialect.SQL92DatabaseType;
+import org.apache.shardingsphere.infra.metadata.schema.builder.loader.dialect.AbstractDatabaseMetaDataDialectHandlerTest;
 import org.apache.shardingsphere.sql.parser.sql.common.constant.QuoteCharacter;
 import org.junit.Test;
 
@@ -28,25 +28,25 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.when;
 
-public final class H2DatabaseMetaDataDialectHandlerTest extends AbstractDatabaseMetaDataDialectHandlerTest {
+public final class SQL92DatabaseMetaDataDialectHandlerTest extends AbstractDatabaseMetaDataDialectHandlerTest {
     
     @Test
     public void assertGetSchema() throws SQLException {
         when(getConnection().getSchema()).thenReturn(DATABASE_NAME);
-        String h2Schema = getSchema(new H2DatabaseType());
-        assertThat(h2Schema, is(DATABASE_NAME));
+        String sql92Schema = getSchema(new SQL92DatabaseType());
+        assertThat(sql92Schema, is(DATABASE_NAME));
     }
     
     @Test
     public void assertFormatTableNamePattern() {
-        String h2TableNamePattern = formatTableNamePattern(new H2DatabaseType());
-        assertThat(h2TableNamePattern, is(TABLE_NAME_PATTERN));
+        String sql92TableNamePattern = formatTableNamePattern(new SQL92DatabaseType());
+        assertThat(sql92TableNamePattern, is(TABLE_NAME_PATTERN));
     }
     
     @Test
     public void assertGetQuoteCharacter() {
-        QuoteCharacter h2QuoteCharacter = getQuoteCharacter(new H2DatabaseType());
-        assertThat(h2QuoteCharacter.getStartDelimiter(), is("\""));
-        assertThat(h2QuoteCharacter.getEndDelimiter(), is("\""));
+        QuoteCharacter sql92QuoteCharacter = getQuoteCharacter(new SQL92DatabaseType());
+        assertThat(sql92QuoteCharacter.getStartDelimiter(), is("\""));
+        assertThat(sql92QuoteCharacter.getEndDelimiter(), is("\""));
     }
 }
